+++ src/backend/src/main/java/com/becon/opencelium/backend/resource/OidcTicketDTO.java
@@ -1,0 +1,4 @@
+package com.becon.opencelium.backend.resource;
+
+public record OidcTicketDTO(String code) {
+}